backend/pyproject.toml
[project]
name = "official_proj"
version = "0.1.0"
description = "ai写作平台 using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.18.0",
    "crewai[tools]==1.8.0",
    "fastapi>=0.128.0",
    "python-jose[cryptography]>=3.3.0",
    "loguru>=0.7.3",
    "ningfastforge>=0.1.8.2",
    "pymilvus>=2.6.6",
    "pymongo>=4.16.0",
    "pymysql>=1.1.2",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
official_proj = "official_proj.main:run"
run_crew = "official_proj.main:run"
train = "official_proj.main:train"
replay = "official_proj.main:replay"
test = "official_proj.main:test"
run_with_trigger = "official_proj.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
